+++ build.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-# Clean install
-pnpm run clean
-pnpm install
-
-# Build
-pnpm run build